feature/2: adjust test DB config s.t. it is ready for REST test - load data model DDL from setup script instead from test suite
