+++ jest.dbTest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  testRegex: '.+dbTest\\.js',
+  testRegex: '.+\\.(db)?[Tt]est\\.js',
   globalSetup: '<rootDir>tests/db/testDbSetup.js',
   globalTeardown: '<rootDir>tests/db/testDbTeardown.js',
 }
